+++ src/common.rs
@@ -1,11 +1,16 @@
+mod interface;
+mod services;
+
 use anyhow::{anyhow, Result};
+pub use interface::{Nic, NicOutput};
 use serde::{Deserialize, Serialize};
-use std::{collections::HashMap, fmt, process::Command};
+pub use services::waitfor_up;
+use std::process::Command;
 
 pub const DEFAULT_PATH_ENV: &str = "/usr/sbin:/usr/bin:/sbin:/bin:/usr/local/aice/bin";
 
 /// Types of command to node.
-#[derive(Debug, Deserialize, Serialize, Clone)]
+#[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
 pub enum Node {
     Hostname(SubCommand),
     Interface(SubCommand),
@@ -46,7 +51,7 @@ impl NodeRequest {
     }
 }
 
-#[derive(Copy, Clone, Debug, Deserialize, Serialize)]
+#[derive(Copy, Clone, Debug, Deserialize, Serialize, PartialEq)]
 pub enum SubCommand {
     Add,
     Delete,
@@ -60,110 +65,6 @@ pub enum SubCommand {
     SetProductVersion,
     Status,
     Update,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct Nic {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub addresses: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub dhcp4: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub gateway4: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub nameservers: Option<HashMap<String, Vec<String>>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub optional: Option<bool>,
-}
-
-impl fmt::Display for Nic {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if let Ok(s) = serde_yaml::to_string(self) {
-            write!(f, "{s}")
-        } else {
-            Ok(())
-        }
-    }
-}
-
-impl Nic {
-    #[must_use]
-    pub fn new(
-        addresses: Option<Vec<String>>,
-        dhcp4: Option<bool>,
-        gateway4: Option<String>,
-        nameservers: Option<HashMap<String, Vec<String>>>,
-        optional: Option<bool>,
-    ) -> Self {
-        Nic {
-            addresses,
-            dhcp4,
-            gateway4,
-            nameservers,
-            optional,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct NicOutput {
-    pub addresses: Option<Vec<String>>,
-    pub dhcp4: Option<bool>,
-    pub gateway4: Option<String>,
-    pub nameservers: Option<Vec<String>>,
-}
-
-impl NicOutput {
-    #[must_use]
-    pub fn new(
-        addresses: Option<Vec<String>>,
-        dhcp4: Option<bool>,
-        gateway4: Option<String>,
-        nameservers: Option<Vec<String>>,
-    ) -> Self {
-        NicOutput {
-            addresses,
-            dhcp4,
-            gateway4,
-            nameservers,
-        }
-    }
-
-    #[must_use]
-    pub fn to(&self) -> Nic {
-        let nameservers = if let Some(nm) = &self.nameservers {
-            let mut m = HashMap::new();
-            m.insert("addresses".to_string(), nm.clone());
-            m.insert("search".to_string(), Vec::new());
-            Some(m)
-        } else {
-            None
-        };
-        Nic {
-            addresses: self.addresses.clone(),
-            dhcp4: self.dhcp4,
-            gateway4: self.gateway4.clone(),
-            nameservers,
-            optional: None,
-        }
-    }
-
-    #[must_use]
-    pub fn from(nic: &Nic) -> Self {
-        let nameservers = {
-            if let Some(nm) = &nic.nameservers {
-                nm.get("addresses").cloned()
-            } else {
-                None
-            }
-        };
-        NicOutput {
-            addresses: nic.addresses.clone(),
-            dhcp4: nic.dhcp4,
-            gateway4: nic.gateway4.clone(),
-            nameservers,
-        }
-    }
 }
 
 /// Runs a linux command and returns its output.