+++ src/root.rs
@@ -7,11 +7,15 @@ mod syslog;
 pub(crate) mod task;
 mod ufw;
 
-use super::common::{run_command_output, Nic, NicOutput, SubCommand, DEFAULT_PATH_ENV};
+use super::common::{run_command_output, Nic, NicOutput, SubCommand};
 use anyhow::{anyhow, Result};
+use roxy::common::DEFAULT_PATH_ENV;
 use std::process::Command;
 
-fn run_command(cmd: &str, path: Option<&[&str]>, args: &[&str]) -> Result<bool> {
+/// Run linux command without output string.
+/// # Errors
+/// * errors are found when execute command or read the execution status.
+pub fn run_command(cmd: &str, path: Option<&[&str]>, args: &[&str]) -> Result<bool> {
     let mut cmd = Command::new(cmd);
     let val = if let Some(path) = path {
         let mut temp = DEFAULT_PATH_ENV.to_string();